+++ package.json
@@ -27,8 +27,7 @@
   "scripts": {
     "dev": "npm start -w insomnia",
     "dev:autoRestart": "npm run start:autoRestart -w insomnia",
-    "check:renderer-node-imports": "npm run check:renderer-node-imports -w insomnia",
-    "lint": "npm run lint --workspaces --if-present",
+"lint": "npm run lint --workspaces --if-present",
     "type-check": "npm run type-check --workspaces --if-present",
     "test": "npm run test --workspaces --if-present",
     "clean": "git clean -dfX",
