+++ packages/insomnia-testing/src/generate/index.ts
@@ -1,5 +1,6 @@
 import type { Test, TestSuite } from './generate';
 
-export { generate, generateToFile } from './generate';
+export { generate } from './generate';
+export { generateToFile } from './generate-to-file';
 
 export type { Test, TestSuite };
