+++ packages/insomnia-testing/src/index.ts
@@ -1,6 +1,7 @@
 import type { Test, TestSuite } from './generate';
 import type { TestResults } from './run';
-export { generate, generateToFile } from './generate';
+export { generate } from './generate';
+export { generateToFile } from './generate/generate-to-file';
 
 export { runTests, runTestsCli } from './run';
