--- packages/insomnia/config/renderer-node-import-baseline.json
@@ -1,56 +1,0 @@
-{
-  "entries": [
-    {
-      "importer": "../insomnia-testing/src/generate/generate.ts",
-      "builtin": "fs"
-    },
-    {
-      "importer": "../insomnia-testing/src/run/run.ts",
-      "builtin": "fs"
-    },
-    {
-      "importer": "../insomnia-testing/src/run/run.ts",
-      "builtin": "os"
-    },
-    {
-      "importer": "../insomnia-testing/src/run/run.ts",
-      "builtin": "path"
-    },
-    {
-      "importer": "src/plugins/context/response.ts",
-      "builtin": "fs"
-    },
-    {
-      "importer": "src/plugins/context/response.ts",
-      "builtin": "zlib"
-    },
-    {
-      "importer": "src/plugins/index.ts",
-      "builtin": "fs"
-    },
-    {
-      "importer": "src/plugins/index.ts",
-      "builtin": "path"
-    },
-    {
-      "importer": "src/scripting/require-interceptor.ts",
-      "builtin": "buffer"
-    },
-    {
-      "importer": "src/scripting/require-interceptor.ts",
-      "builtin": "timers"
-    },
-    {
-      "importer": "src/scripting/require-interceptor.ts",
-      "builtin": "util"
-    },
-    {
-      "importer": "src/templating/liquid-extension.ts",
-      "builtin": "crypto"
-    },
-    {
-      "importer": "src/templating/liquid-extension.ts",
-      "builtin": "os"
-    }
-  ]
-}
